+++ src/lw_common/parse_config/syntax_to_lw_syntax.cs
@@ -5,7 +5,7 @@ using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace lw_common.parse_syntax
+namespace lw_common
 {
     internal class syntax_to_lw_syntax {
         private static log4net.ILog logger = log4net.LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);