+++ src/lw_common/util/db_util.cs
@@ -34,6 +34,45 @@ namespace lw_common
             return null;
         }
 
+        // example:
+        //  <connectionType value="System.Data.SqlClient.SqlConnection, System.Data, Version=1.0.3300.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
+        public static string db_connection_type_to_db_provider(string connection_type) {
+            connection_type = connection_type.ToLower();
+            if (connection_type.Contains(".sqlclient"))
+                return "System.Data.SqlClient";
+            if (connection_type.Contains(".oledb"))
+                return "System.Data.OleDb";
+            if (connection_type.Contains(".odbc"))
+                return "System.Data.Odbc";
+            if (connection_type.Contains(".oracleclient"))
+                return "System.Data.OracleClient";
+            if (connection_type.Contains(".sqlite"))
+                return "System.Data.SQLite";
+
+            return "";
+        }
+
+        private static string to_db_field_name(string name) {
+            name = name.Trim().ToLower();
+            if (name.StartsWith("[") && name.EndsWith("]"))
+                name = name.Substring(1, name.Length - 2);
+            return name;
+        }
+        public static List<string> insert_into_to_db_fields(string sql) {
+            List<string> fields = new List<string>();
+            if (sql.StartsWith("insert into")) {
+                sql = sql.Substring(11).Trim();
+                int separator = sql.IndexOf("(");
+                int end_separator = sql.IndexOf(")");
+                if ( separator > 0 && end_separator > 0) {
+                    ++separator;
+                    string fields_str = sql.Substring(separator, end_separator - separator);
+                    fields = fields_str.Split(new[] {","}, StringSplitOptions.RemoveEmptyEntries).Select(to_db_field_name).ToList();
+                }
+            }
+            return fields;
+        } 
+
         public static string non_null_str(DbDataReader rs, int idx) {
             return rs.IsDBNull(idx) ? "" : rs.GetString(idx);
         }